feat(buy): adjusted buy route and abstracted logic to hooks
